% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planet.R
\name{aberration_corrected_coordinates_planet}
\alias{aberration_corrected_coordinates_planet}
\title{Planetary aberration correction (effect of light-time and Earth's motion)}
\usage{
aberration_corrected_coordinates_planet(jd, planet)
}
\arguments{
\item{jd}{Julian day}

\item{planet}{planet name as string: mercury, venus,earth, mars, jupyter, saturn}
}
\value{
\itemize{
  \item d Corrected distance to Earth
  \item x Corrected rectangular geocentric coordinate of the planet x
  \item y Corrected rectangular geocentric coordinate y
  \item z Corrected rectangular geocentric coordinate z
  \item L Corrected heliocentric longitude of the planet
  \item B Corrected heliocentric latitude of the planet
  \item R Corrected heliocentric radius of the planet

}
}
\description{
Planetary aberration correction (effect of light-time and Earth's motion)
}
\examples{
aberration_corrected_coordinates_planet(julian_day(13.19,11,2028),"venus")

}
