% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.R
\name{aberration_correction_star}
\alias{aberration_correction_star}
\title{Annual aberration correction}
\usage{
aberration_correction_star(jd, declination, right_ascension,
  precise = TRUE)
}
\arguments{
\item{jd}{Julian day}

\item{declination}{Declination in degrees}

\item{right_ascension}{Right ascension in hour angle}

\item{precise}{Precise calculation}
}
\value{
\itemize{
  \item declination correction in declination due to annual aberration
  \item right_ascension correction in right ascension due to annual aberration
}
}
\description{
Annual aberration correction
}
\examples{
declination<-str2degrees("19รยบ10'57\\u0022")
right_ascension<-str2hours("14:15:39.7")
jd<-julian_day(13.19,11,2028)
aberration_correction_star(jd,declination,right_ascension)

}
