% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.R
\name{angular_separation}
\alias{angular_separation}
\title{Angular separation between two celestial bodies}
\usage{
angular_separation(declination1, right_ascension1, declination2,
  right_ascension2)
}
\arguments{
\item{declination1}{Declination of first celestial body in degrees}

\item{right_ascension1}{Right ascension of first celestial body in hour angle}

\item{declination2}{Declination of second celestial body in degrees}

\item{right_ascension2}{Right ascension of second celestial body in hour angle}
}
\value{
Angular separation
}
\description{
Angular separation between two celestial bodies
}
\examples{
declination1<-str2degrees("19º10'57\\u0022")
right_ascension1<-str2hours("14:15:39.7")
declination2<-str2degrees("-11º09'41\\u0022")
right_ascension2<-str2hours("13:25:11.6")
angular_separation(declination1,right_ascension1, declination2, right_ascension2)

}
