% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roots.R
\name{apparent_elevation}
\alias{apparent_elevation}
\title{Function for getting the difference between the apparent elevation coordinate according to some parameters and an reference elevation}
\usage{
apparent_elevation(x, parms)
}
\arguments{
\item{x}{Julian day}

\item{parms}{Vector of strings containing the name of celestial body, geographic longitude, geographic latitude, geographic altitude and reference elevation}
}
\value{
Difference in elevation
}
\description{
Function for getting the difference between the apparent elevation coordinate according to some parameters and an reference elevation
}
