% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{apparent_equatorial_position}
\alias{apparent_equatorial_position}
\title{Apparent equatorial coordinates of a celestial body}
\usage{
apparent_equatorial_position(jd, celestial_body)
}
\arguments{
\item{jd}{Julian day}

\item{celestial_body}{celestial_body name as string: mercury, venus, mars, jupyter, saturn, moon, sun...}
}
\value{
Equatorial coordinates
\itemize{
  \item declination Declination in degrees
  \item right_ascension Right ascension in hour angle
  \item distance Distance to Earth (planet or moon)
  \item illuminated_fraction Illuminated fraction of the disk of a planet  or moon (planet or moon)
  \item apparent_magnitude apparent magnitude (planet or moon)
}#'
}
\description{
Apparent equatorial coordinates of a celestial body
}
\examples{
jd<-julian_day(13.19,11,2028)
apparent_equatorial_position_planet(jd,"venus")

}
