% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planet.R
\name{apparent_equatorial_position_planet}
\alias{apparent_equatorial_position_planet}
\title{Apparent equatorial coordinates of a planet}
\usage{
apparent_equatorial_position_planet(jd, planet)
}
\arguments{
\item{jd}{Julian day}

\item{planet}{planet name as string: mercury, venus,earth, mars, jupyter, saturn}
}
\value{
Equatorial coordinates
\itemize{
  \item declination Declination in degrees
  \item right_ascension Right ascension in hour angle
  \item distance Distance to Earth
  \item illuminated_fraction Illuminated fraction of the disk of a planet
  \item apparent_magnitude apparent magnitude
}#'
}
\description{
Apparent equatorial coordinates of a planet
}
\examples{
jd<-julian_day(13.19,11,2028)
apparent_equatorial_position_planet(jd,"venus")

}
