% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.R
\name{apparent_equatorial_position_star}
\alias{apparent_equatorial_position_star}
\title{Equatorial coordinates for a star}
\usage{
apparent_equatorial_position_star(jd,
  former_jd = str2jd("2000-01-01T12:00"), former_declination,
  former_right_ascension, proper_motion_declination,
  proper_motion_right_ascension, precise = TRUE)
}
\arguments{
\item{jd}{Julian day}

\item{former_jd}{Julian day for reference (usually J2000)}

\item{former_declination}{Reference declination in degrees}

\item{former_right_ascension}{Reference right ascension in hour angle}

\item{proper_motion_declination}{Proper motion declination in degrees}

\item{proper_motion_right_ascension}{Proper motion right ascension in hour angle}

\item{precise}{if precise calculation}
}
\value{
Equatorial coordinates
\itemize{
  \item declination Declination in degrees
  \item right_ascension Right ascension in hour angle
}
}
\description{
Equatorial coordinates for a star
}
\examples{
jd<-julian_day(13.19,11,2028)
former_jd<-str2jd("2000-01-01T12:00")
fd<-str2degrees("49รยบ13'42.48\\u0022")
fra<-str2hours("02:44:11.986")
apparent_equatorial_position_star(jd,former_jd,fd,fra, -0.0895/3600, 0.03425/3600)

}
