% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.R
\name{apparent_equatorial_position_sun}
\alias{apparent_equatorial_position_sun}
\title{Apparent equatorial coordinates of sun}
\usage{
apparent_equatorial_position_sun(jd, precise = TRUE)
}
\arguments{
\item{jd}{Julian day}

\item{precise}{Precise calculation}
}
\value{
Equatorial coordinates
\itemize{
  \item declination Declination in degrees
  \item right_ascension Right ascension in hour angle
}#'
}
\description{
Apparent equatorial coordinates of sun
}
\examples{
jd<-julian_day(13.19,11,2028)
apparent_longitude_sun(jd)

}
