% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidereal_time_greenwich.R
\name{apparent_sidereal_time_at_Greenwich}
\alias{apparent_sidereal_time_at_Greenwich}
\title{Apparent sidereal time at Greenwich}
\usage{
apparent_sidereal_time_at_Greenwich(jd)
}
\arguments{
\item{jd}{Julian day}
}
\value{
Julian day
}
\description{
Apparent sidereal time at Greenwich
}
\examples{
mean_sidereal_time_at_Greenwich(c(2436116.31, 1842713.00))

}
