% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moon.R
\name{argument_of_latitude_moon}
\alias{argument_of_latitude_moon}
\title{Moon`s argument of latitude (mean distance of the Moon from its ascending node) referred to the mean equinox of the date}
\usage{
argument_of_latitude_moon(jd)
}
\arguments{
\item{jd}{Julian day}
}
\value{
Moon`s argument of latitude in degrees
}
\description{
Moon`s argument of latitude (mean distance of the Moon from its ascending node) referred to the mean equinox of the date
}
\examples{
jd<-julian_day(13.19,11,2028)
argument_of_latitude_moon(jd)

}
