% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roots.R
\name{azimuth}
\alias{azimuth}
\title{Function for getting the difference between the azimuth coordinate according to some parameters and an reference azimuth}
\usage{
azimuth(x, parms)
}
\arguments{
\item{x}{Julian day}

\item{parms}{Vector of strings containing the name of celestial body, geographic longitude, geographic latitude, geographic altitude  and reference azimuth}
}
\value{
Difference in azimuth
}
\description{
Function for getting the difference between the azimuth coordinate according to some parameters and an reference azimuth
}
\examples{
from <- str2jd("2014-03-08")
to <- str2jd("2015-03-08")
azimuth(seq(from,to,1),parms=c("sun","42°36'25\\"N","5°33'58\\"W","782", "20°0'0\\""))

}
