% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidereal_time_greenwich.R
\name{centuries_since_2000}
\alias{centuries_since_2000}
\title{Number of centuries since 2000-01-01}
\usage{
centuries_since_2000(jd)
}
\arguments{
\item{jd}{Julian day}
}
\value{
Number of centuries since 2000-01-01
}
\description{
Number of centuries since 2000-01-01
}
\examples{
centuries_since_2000(c(2436116.31, 1842713.00))

}
