% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{check_alignment}
\alias{check_alignment}
\title{Get error in azimuth given the horizontal coordinates of an alignment in reference to a theorical declination}
\usage{
check_alignment(azimuth, altitude, celestial_body, jd = julian_day(1, 1,
  -1000), disk, declination_reference, geographic_latitude)
}
\arguments{
\item{azimuth}{Apparent azimuth in degrees of an alignment}

\item{altitude}{Apparent elevation in degrees of an alignment}

\item{celestial_body}{celestial body as string. It can be empty}

\item{jd}{Julian day. -1000-1-1 as default}

\item{disk}{It can be "upper" for upper limb, "lower" for lower limb or "center"}

\item{declination_reference}{Theorical declination as reference}

\item{geographic_latitude}{Geographical latitude of the origin in degrees}
}
\value{
Error in azimuth
}
\description{
Get error in azimuth given the horizontal coordinates of an alignment in reference to a theorical declination
}
\examples{
df<- data.frame(origin=c("Fuencalada","Fuencalada"), foresight=c("Los Cambitos","Teleno"),
                azimuth=c(241.4563,246.9905),altitude=c(2.035853,2.192668),
                geographic_latitude=c(42.44944,42.44944))
error=check_alignment(azimuth=df$azimuth,altitude=df$altitude,
                      celestial_body=c("moon","sun"),
                      disk=c("upper","upper"),
                      declination_reference=c(-18.81,-16.4),
                      geographic_latitude=df$geographic_latitude)
}
