% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heliocentric.R
\name{cos_component}
\alias{cos_component}
\title{Cosine auxiliary function}
\usage{
cos_component(A, B, C, t)
}
\arguments{
\item{A}{Parameter A}

\item{B}{Parameter B}

\item{C}{Parameter C}

\item{t}{Parameter t}
}
\value{
A * cos(B + C * t)
}
\description{
Cosine auxiliary function
}
