% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roots.R
\name{crossing_zero_up}
\alias{crossing_zero_up}
\title{Zero crossing of a periodical function}
\usage{
crossing_zero_up(fnc, from, to, period, search_interval = 0.005,
  incrx = 0.01, silent = FALSE, trace = TRUE)
}
\arguments{
\item{fnc}{Function}

\item{from}{Initial value of the interval to be searched}

\item{to}{Final value of the interval to be searched}

\item{period}{Period of the function}

\item{search_interval}{Search interval for each root}

\item{incrx}{increment in x to calculate the derivative used to check the slope in the point of crossing}

\item{silent}{If true a text progress bar is shown}

\item{trace}{If true prints a trace}
}
\value{
vector of roots of the function
}
\description{
Zero crossing of a periodical function
}
\examples{
fnc<-function(x) apparent_equatorial_position(x,"sun")$declination
from <- str2jd("2014-03-08")
to <- str2jd("2018-03-08")
crossing_zero_up(fnc,from,to,period=365.25,search_interval=10)

}
