% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{degrees2hours}
\alias{degrees2hours}
\title{Degrees to hour angle transformation}
\usage{
degrees2hours(degrees)
}
\arguments{
\item{degrees}{Degrees}
}
\value{
Hour angle
}
\description{
Degrees to hour angle transformation
}
\examples{
degrees2hours(180)

}
