% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{degrees2rad}
\alias{degrees2rad}
\title{Degrees to radian transformation}
\usage{
degrees2rad(degrees)
}
\arguments{
\item{degrees}{Degrees}
}
\value{
Radians
}
\description{
Degrees to radian transformation
}
\examples{
degrees2rad(180)

}
