% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{degrees2str}
\alias{degrees2str}
\title{Sexagesimal degree formatting to string}
\usage{
degrees2str(degrees)
}
\arguments{
\item{degrees}{Degrees}
}
\value{
string d°m's" formatted
}
\description{
It converts angle into string
}
\examples{
degrees2str(c(15.5,30.14))

}
