% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roots.R
\name{derivative}
\alias{derivative}
\title{First derivative for the function and value given}
\usage{
derivative(fnc, x, incrx = 0.01)
}
\arguments{
\item{fnc}{Function}

\item{x}{Value in which the derivative should be calculated}

\item{incrx}{increment in x to calculate the derivative}
}
\value{
derivative value at the value given
}
\description{
First derivative for the function and value given
}
\examples{
fnc<-function(x) apparent_equatorial_position(x,"sun")$declination
x <- str2jd("2014-03-08")
derivative(fnc,x)

}
