% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{distance_to_earth}
\alias{distance_to_earth}
\title{Distance to Earth with planetary aberration correction (effect of light-time and Earth's motion)}
\usage{
distance_to_earth(jd, celestial_body)
}
\arguments{
\item{jd}{Julian day}

\item{celestial_body}{celestial_body name as string: mercury, venus, mars, jupyter, saturn, moon, sun...}
}
\value{
distance to Earth
}
\description{
Distance to Earth with planetary aberration correction (effect of light-time and Earth's motion)
}
\examples{
distance_to_earth(julian_day(13.19,11,2028),"venus")

}
