% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moon.R
\name{distance_to_earth_moon}
\alias{distance_to_earth_moon}
\title{Distance to Earth}
\usage{
distance_to_earth_moon(jd)
}
\arguments{
\item{jd}{Julian day}
}
\value{
Distance to Earth
}
\description{
Distance to Earth
}
\examples{
jd<-julian_day(13.19,11,2028)
distance_to_earth_moon(jd)

}
