% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{dms}
\alias{dms}
\title{Sexagesimal degre components}
\usage{
dms(degrees)
}
\arguments{
\item{degrees}{Degrees}
}
\value{
sexagesimal degree components
\itemize{
  \item degrees
  \item minutes
  \item seconds
}
}
\description{
It retrieves the sexagesimal degree components, i.e. degrees, minutes and seconds
}
\examples{
dms(c(15.5,30.14))

}
