% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian_day.R
\name{dmyh}
\alias{dmyh}
\title{Julian day components: day, month, year and hour}
\usage{
dmyh(jd)
}
\arguments{
\item{jd}{Julian day}
}
\value{
Julian day components
\itemize{
  \item day
  \item month
  \item year
  \item hour from 0 to 24
}
}
\description{
Julian day components: day, month, year and hour
}
\examples{
dmyh(2443259.9)

}
