% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{ecliptic2equatorial}
\alias{ecliptic2equatorial}
\title{Ecliptic coordinates to  Equatorial coordinates transformation}
\usage{
ecliptic2equatorial(longitude, latitude, jd = NULL, obliquity = NULL)
}
\arguments{
\item{longitude}{Longitude in degrees}

\item{latitude}{Latitude in degrees}

\item{jd}{Julian day used to determine true obliquity of ecliptic}

\item{obliquity}{true obliquity of ecliptic. If null it uses jd to calculate it}
}
\value{
Equatorial coordinates
\itemize{
  \item declination Declination in degrees
  \item right_ascension Right ascension in hour angle
}
}
\description{
Ecliptic coordinates to  Equatorial coordinates transformation
}
\examples{
ecliptic2equatorial(longitude = 113.2156292276,latitude=6.684170072045,obliquity=23.4392911)

}
