% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{equatorial2ecliptic}
\alias{equatorial2ecliptic}
\title{Equatorial coordinates to Ecliptic coordinates transformation}
\usage{
equatorial2ecliptic(declination, right_ascension, jd = NULL,
  obliquity = NULL)
}
\arguments{
\item{declination}{Declination in degrees}

\item{right_ascension}{Right ascension in hour angle}

\item{jd}{Julian day used to determine true obliquity of ecliptic}

\item{obliquity}{true obliquity of ecliptic. If null it uses jd to calculate it}
}
\value{
Ecliptic coordinates
\itemize{
  \item latitude
  \item longitude longitude in degrees
}
}
\description{
Equatorial coordinates to Ecliptic coordinates transformation
}
