% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{equatorial2horizontal}
\alias{equatorial2horizontal}
\title{Equatorial coordinates to  Horizontal coordinates transformation}
\usage{
equatorial2horizontal(geographic_longitude, geographic_latitude,
  declination, right_ascension, jd = NULL, obliquity = NULL)
}
\arguments{
\item{geographic_longitude}{Geographical longitude in degrees}

\item{geographic_latitude}{Geographical latitude in degrees}

\item{declination}{Declination in degrees}

\item{right_ascension}{Right Ascension as hour angle}

\item{jd}{Julian day used to determine true obliquity of ecliptic}

\item{obliquity}{true obliquity of ecliptic. If null it uses jd to calculate it}
}
\value{
Horizontal coordinates
\itemize{
  \item azimuth Azimuth in degrees
  \item altitude Altitude in hour angle
}
}
\description{
Equatorial coordinates to  Horizontal coordinates transformation
}
\examples{
geographic_longitude<-str2degrees("-77º03'55.5\\u0022")
geographic_latitude<-str2degrees("38º55'17\\u0022")
jd<-str2jd("1987-04-10T19:21:00")
declination<-str2degrees("-6º43'11.61\\u0022")
right_ascension<-str2hours("23:09:16.641")
equatorial2horizontal(geographic_longitude,geographic_latitude,declination, right_ascension,jd)
}
