% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{extinction_angle}
\alias{extinction_angle}
\title{Extinction zenith angle}
\usage{
extinction_angle(star_visual_magnitude,
  visual_extinction_coefficient = 0.2)
}
\arguments{
\item{star_visual_magnitude}{Visual magnitude of a star}

\item{visual_extinction_coefficient}{Visual extinction coefficient}
}
\value{
Extinction angle, in degrees
}
\description{
Extinction zenith angle
}
\examples{
sky_brightness(degrees2rad(150),degrees2rad(-10),degrees2rad(88))

}
