% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{festivals}
\alias{festivals}
\title{Dates of solstices, equinoxes and cross-quarter days in an interval}
\usage{
festivals(from, to, silent = FALSE)
}
\arguments{
\item{from}{Initial value of the interval to be searched}

\item{to}{Final value of the interval to be searched}

\item{silent}{If false a text progress bar is shown}
}
\value{
a data frame with two columns: festival identifying the festival and date the corresponding Julian date
}
\description{
Dates of solstices, equinoxes and cross-quarter days in an interval
}
\examples{
from <- str2jd("2014-03-08")
to <- str2jd("2018-03-08")
festivals(from,to)

}
