% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{first_new_moon}
\alias{first_new_moon}
\title{Dates of first visibility of the lunar crescent just after a solar festival (solstice, equinox or cross-quarter day)}
\usage{
first_new_moon(from, to, festivities = NULL,
  festival_reference = "winter solstice", geographical_latitude,
  geographical_longitude, geographical_altitude, silent = FALSE)
}
\arguments{
\item{from}{Initial value of the interval to be searched}

\item{to}{Final value of the interval to be searched}

\item{festivities}{Festivities data frame obtained with festivals() function}

\item{festival_reference}{Solar festival (solstice, equinox or cross-quarter day) as reference for the first new moon}

\item{geographical_latitude}{Geographical latitude in degrees}

\item{geographical_longitude}{Geographical longitude in degrees}

\item{geographical_altitude}{Geographical altitude in meters}

\item{silent}{If false a text progress bar is shown}
}
\value{
Dates of first visibility of the lunar crescent
}
\description{
Dates of first visibility of the lunar crescent just after a solar festival (solstice, equinox or cross-quarter day)
}
