% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{geocentric2topocentric}
\alias{geocentric2topocentric}
\title{Geocentric to topocentric equatorial coordinates transformation}
\usage{
geocentric2topocentric(declination, right_ascension, jd,
  geographic_longitude, geographic_latitude, geographic_altitude,
  celestial_body)
}
\arguments{
\item{declination}{Geocentric declination in degrees}

\item{right_ascension}{Geocentric right Ascension in hour angle}

\item{jd}{Julian day used to determine true obliquity of ecliptic}

\item{geographic_longitude}{Geographical longitude in degrees}

\item{geographic_latitude}{Geographical latitude in degrees}

\item{geographic_altitude}{Geographical altitude in meters}

\item{celestial_body}{Celestial body name as string: mercury, venus,mars, jupyter, saturn, sun or moon}
}
\value{
Horizontal coordinates
\itemize{
  \item declination Topocentric declination in degrees
  \item right_ascension Topocentric right ascension in hour angle
}
}
\description{
Geocentric to topocentric equatorial coordinates transformation
}
\examples{
jd<-str2jd("2003-08-28T03:17:00")
geocentric2topocentric(-15.771083,-22.63534722222,jd,-116.8625,-33.356111,1706,"mars")

}
