% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_planets}
\alias{get_planets}
\title{Return the list of planets that can be seen withe the naked eye}
\usage{
get_planets()
}
\value{
list of the stars included in the catalogue
}
\description{
Return the list of planets that can be seen withe the naked eye
}
\examples{
get_stars()

}
