% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{get_stars}
\alias{get_stars}
\title{Return a list of the stars included in the catalogue}
\usage{
get_stars()
}
\value{
list of the stars included in the catalogue
}
\description{
Return a list of the stars included in the catalogue
}
\examples{
get_stars()

}
