% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heliocentric.R
\name{heliocentric_latitude}
\alias{heliocentric_latitude}
\title{Heliocentric latitude of a planet}
\usage{
heliocentric_latitude(jd, celestial_body)
}
\arguments{
\item{jd}{Julian day}

\item{celestial_body}{planet name as string: mercury, venus,earth, mars, jupyter, saturn}
}
\value{
Heliocentric longitude  in degrees
}
\description{
Heliocentric latitude of a planet
}
\examples{
jd<-julian_day(13.19,11,2028)
heliocentric_latitude(jd,"venus")

}
