% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heliocentric.R
\name{heliocentric_radius}
\alias{heliocentric_radius}
\title{Heliocentric radius of a planet}
\usage{
heliocentric_radius(jd, celestial_body)
}
\arguments{
\item{jd}{Julian day}

\item{celestial_body}{planet name as string: mercury, venus,earth, mars, jupyter, saturn}
}
\value{
Heliocentric radius
}
\description{
Heliocentric radius of a planet
}
\examples{
jd<-julian_day(13.19,11,2028)
heliocentric_radius(jd,"venus")

}
