% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian_day.R
\name{hms}
\alias{hms}
\title{Hour components: hours, minutes, seconds}
\usage{
hms(hour_value)
}
\arguments{
\item{hour_value}{Decimal number of hours}
}
\value{
Hour components
\itemize{
  \item hours
  \item minutes
  \item seconds
}
}
\description{
Hour components: hours, minutes, seconds
}
\examples{
hms(c(19.4,12))

}
