% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{horizontal2equatorial}
\alias{horizontal2equatorial}
\title{Equatorial coordinates to  Horizontal coordinates transformation}
\usage{
horizontal2equatorial(geographic_longitude, geographic_latitude, azimuth,
  altitude, jd = NULL, obliquity = NULL)
}
\arguments{
\item{geographic_longitude}{Geographical longitude in degrees}

\item{geographic_latitude}{Geographical latitude in degrees}

\item{azimuth}{Azimuth in degrees}

\item{altitude}{Altitude in degrees}

\item{jd}{Julian day used to determine true obliquity of ecliptic}

\item{obliquity}{true obliquity of ecliptic. If null it uses jd to calculate it}
}
\value{
Horizontal coordinates
\itemize{
  \item azimuth Azimuth in degrees
  \item altitude Altitude in hour angle
}
}
\description{
Equatorial coordinates to  Horizontal coordinates transformation
}
\examples{
geographic_longitude<-str2degrees("-77º03'55.5\\u0022")
geographic_latitude<-str2degrees("38º55'17\\u0022")
jd<-str2jd("1987-04-10T19:21:00")
azimuth<-248.0336941301
altitude<-15.1248736234
horizontal2equatorial(geographic_longitude,geographic_latitude,azimuth, altitude,jd)

}
