% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{hours2degrees}
\alias{hours2degrees}
\title{Hour angle to degrees transformation}
\usage{
hours2degrees(hour_angle)
}
\arguments{
\item{hour_angle}{Hour angle}
}
\value{
Degrees
}
\description{
Hour angle to degrees transformation
}
\examples{
hours2degrees(12)

}
