% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{hours2rad}
\alias{hours2rad}
\title{Hour angle to radian transformation}
\usage{
hours2rad(hour_angle)
}
\arguments{
\item{hour_angle}{Hour angle}
}
\value{
Radians
}
\description{
Hour angle to radian transformation
}
\examples{
hours2rad(12)

}
