% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian_day.R
\name{hours2str}
\alias{hours2str}
\title{Hour value formatted as hh:mm:ss string}
\usage{
hours2str(hour_value)
}
\arguments{
\item{hour_value}{Decimal number of hours}
}
\value{
Hour value formatted as hh:mm:ss string
}
\description{
Hour value formatted as hh:mm:ss string
}
\examples{
reduce_hours(c(19.4,12))

}
