% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian_day.R
\name{jd2str}
\alias{jd2str}
\title{Julian day formatted as IS8601 string}
\usage{
jd2str(jd, short = FALSE)
}
\arguments{
\item{jd}{Julian day}

\item{short}{Short string}
}
\value{
Julian day formatted as IS8601 string
}
\description{
Julian day formatted as IS8601 string
}
\examples{
julian_day(day=26, month=4, year=1977, hours=9.6)

}
