% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian_day.R
\name{julian_day}
\alias{julian_day}
\title{Julian day from day, month, year and hour angle}
\usage{
julian_day(day, month, year, hours = 0)
}
\arguments{
\item{day}{Day}

\item{month}{Month}

\item{year}{Year}

\item{hours}{Hour angle}
}
\value{
Julian day
}
\description{
Julian day from day, month, year and hour angle
}
\examples{
julian_day(day=c(4,27),month=c(10,1),year=c(1957,333),hours=c(19.44,12))

}
