% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moon.R
\name{mean_anomaly_moon}
\alias{mean_anomaly_moon}
\title{Mean anomaly of the moon referred to the mean equinox of the date}
\usage{
mean_anomaly_moon(jd)
}
\arguments{
\item{jd}{Julian day}
}
\value{
Mean anomaly of the moon in degrees
}
\description{
Mean anomaly of the moon referred to the mean equinox of the date
}
\examples{
jd<-julian_day(13.19,11,2028)
mean_elongation_moon(jd)

}
