% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moon.R
\name{mean_longitude_moon}
\alias{mean_longitude_moon}
\title{Geometric mean longitude of the moon referred to the mean equinox of the date}
\usage{
mean_longitude_moon(jd)
}
\arguments{
\item{jd}{Julian day}
}
\value{
Geometric mean longitude of teh moon in degrees
}
\description{
Geometric mean longitude of the moon referred to the mean equinox of the date
}
\examples{
jd<-julian_day(13.19,11,2028)
mean_longitude_moon(jd)

}
