% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidereal_time_greenwich.R
\name{mean_sidereal_time_at_Greenwich}
\alias{mean_sidereal_time_at_Greenwich}
\title{Mean sidereal time at Greenwich}
\usage{
mean_sidereal_time_at_Greenwich(jd)
}
\arguments{
\item{jd}{Julian day}
}
\value{
Mean sidereal time at Greenwich as hour angle
}
\description{
Mean sidereal time at Greenwich
}
\examples{
mean_sidereal_time_at_Greenwich(c(2436116.31, 1842713.00))

}
