% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sidereal_time_greenwich.R
\name{mean_sidereal_time_at_Greenwich_at_0_UT}
\alias{mean_sidereal_time_at_Greenwich_at_0_UT}
\title{Mean sidereal time at Greenwich at 00:00 UT}
\usage{
mean_sidereal_time_at_Greenwich_at_0_UT(jd)
}
\arguments{
\item{jd}{Julian day}
}
\value{
Mean sidereal time at Greenwich at 00:00 UT as hour angle
}
\description{
Mean sidereal time at Greenwich at 00:00 UT
}
\examples{
mean_sidereal_time_at_Greenwich_at_0_UT(c(2436116.31, 1842713.00))

}
