% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.R
\name{nutation_correction_star}
\alias{nutation_correction_star}
\title{Nutation correction}
\usage{
nutation_correction_star(jd, declination, right_ascension, precise = T)
}
\arguments{
\item{jd}{Julian day}

\item{declination}{Declination in degrees}

\item{right_ascension}{Right ascension in hour angle}

\item{precise}{if precise calculation}
}
\value{
\itemize{
  \item declination correction in declination due to nutation
  \item right_ascension correction in right ascension due to nutation
}
}
\description{
Nutation correction
}
\examples{
declination<-str2degrees("19รยบ10'57\\u0022")
right_ascension<-str2hours("14:15:39.7")
jd<-julian_day(13.19,11,2028)
nutation_correction_star(jd,declination,right_ascension)

}
