% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutation_and_obliquity.R
\name{nutation_in_longitude}
\alias{nutation_in_longitude}
\title{Nutation in longitude}
\usage{
nutation_in_longitude(jd)
}
\arguments{
\item{jd}{Julian day}
}
\value{
Nutation in longitude
}
\description{
Nutation in longitude
}
\examples{
nutation_in_longitude(c(2436116.31, 1842713.00))

}
