% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutation_and_obliquity.R
\name{nutation_in_obliquity}
\alias{nutation_in_obliquity}
\title{Nutation in obliquity}
\usage{
nutation_in_obliquity(jd)
}
\arguments{
\item{jd}{Julian day}
}
\value{
Nutation in obliquity
}
\description{
Nutation in obliquity
}
\examples{
nutation_in_obliquity(c(2436116.31, 1842713.00))

}
