% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/planet.R
\name{planetary_orbit_elements}
\alias{planetary_orbit_elements}
\title{Orbital elements for the mean equinox of the date}
\usage{
planetary_orbit_elements(jd, planet)
}
\arguments{
\item{jd}{Julian day}

\item{planet}{planet name as string: mercury, venus,earth, mars, jupyter, saturn}
}
\value{
\itemize{
  \item mean_longitude Mean longitude of the planet in degrees
  \item semimajor_axis Semimajor axis of the orbit
  \item eccentricity Eccentricity of the orbit
  \item inclination Inclination on the plane of the ecliptic in degrees
  \item ascending_node_longitude Longitude of the ascending node in degrees
  \item perihelion_longitude Longitude of the perihelion in degrees
}
}
\description{
Orbital elements for the mean equinox of the date
}
\examples{
planetary_orbit_elements(julian_day(13.19,11,2028),"venus")

}
