% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.R
\name{precession_star}
\alias{precession_star}
\title{Equatorial coordinates for a stat due to precession}
\usage{
precession_star(jd, former_jd = str2jd("2000-01-01T12:00"),
  former_declination, former_right_ascension, proper_motion_declination,
  proper_motion_right_ascension)
}
\arguments{
\item{jd}{Julian day}

\item{former_jd}{Julian day for reference (usually J2000)}

\item{former_declination}{Reference declination in degrees}

\item{former_right_ascension}{Reference right ascension in hour angle}

\item{proper_motion_declination}{Proper motion declination in degrees}

\item{proper_motion_right_ascension}{Proper motion right ascension in hour angle}
}
\value{
Equatorial coordinates
\itemize{
  \item declination Declination in degrees
  \item right_ascension Right ascension in hour angle
}
}
\description{
Equatorial coordinates for a stat due to precession
}
\examples{
fra<-str2hours("02:44:11.986")
fd<-str2degrees("49รยบ13'42.48\\u0022")
former_jd=str2jd("2000-1-1T12:00")
jd<-julian_day(13.19,11,2028)
pmra<-0.03425/3600
pmd<--0.0895/3600
precession_star(jd, str2jd("2000-01-01T12:00"),fd,fra, pmd, pmra)

}
