% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{probability_fortuitous_alignments.bernoulli}
\alias{probability_fortuitous_alignments.bernoulli}
\title{Fortuitous probability for a number of coincidences for a total of alignments using Bernoulli process}
\usage{
probability_fortuitous_alignments.bernoulli(meaningful = 18,
  coincidences, total, error = 1)
}
\arguments{
\item{meaningful}{Number of meaningful directions. By default 18 are considered: equinoxes, cross-quarter days, solstices, lunar standstills (rising and setting)}

\item{coincidences}{Number of coincidences to within +- error of azimuth}

\item{total}{Number of horizon aligments}

\item{error}{Azimuth error}
}
\value{
Fortuitous probability
}
\description{
Fortuitous probability for a number of coincidences for a total of alignments using Bernoulli process
}
\examples{
probability_fortuitous_alignments.bernoulli(coincidences=3, total=10, error=1)
}
