% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{rad2degrees}
\alias{rad2degrees}
\title{Radian to degrees}
\usage{
rad2degrees(radian)
}
\arguments{
\item{radian}{Radian}
}
\value{
Degrees
}
\description{
Radian to degrees
}
\examples{
rad2degrees(3.141593)

}
