% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{rad2hours}
\alias{rad2hours}
\title{Radian to hour angle}
\usage{
rad2hours(radian)
}
\arguments{
\item{radian}{Radian}
}
\value{
Hour angle
}
\description{
Radian to hour angle
}
\examples{
rad2hours(3.141593)

}
