% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{reduce_degrees}
\alias{reduce_degrees}
\title{Reduce degrees to the interval [0,360), or (-180,180] if signed}
\usage{
reduce_degrees(degrees, signed = FALSE)
}
\arguments{
\item{degrees}{Decimal degrees}

\item{signed}{Signed degrees in the interval (-180,180]}
}
\value{
degrees value reduced to the interval [0,360), or (-180,180] if signed
}
\description{
Reduce degrees to the interval [0,360), or (-180,180] if signed
}
\examples{
reduce_degrees(c(19.4+24*3,12-24*2))

}
