% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian_day.R
\name{reduce_hours}
\alias{reduce_hours}
\title{Reduce hour value to the interval [0,24)}
\usage{
reduce_hours(hour_value)
}
\arguments{
\item{hour_value}{Decimal number of hours}
}
\value{
Hour value reduced to the interval [0,24)
}
\description{
Reduce hour value to the interval [0,24)
}
\examples{
reduce_hours(c(19.4+24*3,12-24*2))

}
