% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{star_visibility}
\alias{star_visibility}
\title{Star visibility information when certain star sets or rises at a given horizontal coordinate value}
\usage{
star_visibility(x, parms, horizontal_coordinate = "elevation",
  horizontal_star_time = "setting")
}
\arguments{
\item{x}{Julian day}

\item{parms}{Vector of strings containing the name of celestial body, geographic longitude, geographic latitude, geographic altitude and reference horizontal coordinate value}

\item{horizontal_coordinate}{Horizontal coordinate (azimuth or elevation) to compare to a reference coordinate value or threshold}

\item{horizontal_star_time}{"setting" or "rising". Only required if horizontal_coordinate=="elevation"}
}
\value{
Star visibility information
\itemize{
  \item date_star Julian day when the star reaches the specified horizontal coordinate value when rising or setting
  \item sky_visual_magnitude Visual magnitude of the sky at the specified horizontal coordinate value when the star is rising or setting
  \item azimuth_sun Azimuth of the Sun when the star reaches the specified horizontal coordinate value when rising or setting
  \item azimuth_star Azimuth of the star when it reaches the specified horizontal  coordinate value when rising or setting
  \item altitude_sun Altitude of the Sun when the star reaches the specified horizontal coordinate value when rising or setting
  \item altitude_star Altitude of the star when it reaches the specified horizontal  coordinate value when rising or setting
  \item star_visual_magnitude Visual magnitude of the star
}#'
}
\description{
Star visibility information when certain star sets or rises at a given horizontal coordinate value
}
\examples{
years<--2000:-1999
from<-str2jd(paste(years[1],"01","01",sep="-"))
to<-str2jd(paste(years[length(years)],"01","01",sep="-"))
star_name<-"epsilon orionis"
ee<-star_visibility(seq(from,to,1),
                    parms=c(star_name,
                            "42\\u00b036'25\\u0022N",
                            "5\\u00b033'58\\u0022W",
                            "782",
                            "2\\u00b00'0\\u0022"),
                    horizontal_coordinate="elevation", horizontal_star_time="setting")

ee<-star_visibility(seq(from,to,1),
                    parms=c(star_name,
                            "42\\u00b036'25\\u0022N",
                            "5\\u00b033'58\\u0022W",
                            "782",
                            "2\\u00b00'0\\u0022"),
                    horizontal_coordinate="elevation", horizontal_star_time="rising")

ee<-star_visibility(seq(from,to,1),
                    parms=c(star_name,
                            "42\\u00b036'25\\u0022N",
                            "5\\u00b033'58\\u0022W",
                            "782",
                            "-107\\u00b022'33.866\\u0022"),
                    horizontal_coordinate="azimuth")

ee<-star_visibility(seq(from,to,1),
                    parms=c(star_name,
                            "42\\u00b036'25\\u0022N",
                            "5\\u00b033'58\\u0022W",
                            "782",
                            "107\\u00b022'33.866\\u0022"),
                    horizontal_coordinate="azimuth")

}
