% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general.R
\name{stellar_phenomena}
\alias{stellar_phenomena}
\title{Julian day for star's appearance or disappearance}
\usage{
stellar_phenomena(from, to, star_name, geographic_latitude,
  geographic_longitude, geographic_altitude, horizontal_coordinate,
  horizontal_star_time, coordinate_threshold,
  visibility_type = "appearance", silent = FALSE)
}
\arguments{
\item{from}{Initial value of the interval to be searched}

\item{to}{Final value of the interval to be searched}

\item{star_name}{Name of the star}

\item{geographic_latitude}{Geographical latitude in degrees}

\item{geographic_longitude}{Geographical longitude in degrees}

\item{geographic_altitude}{Geographical altitude in meters}

\item{horizontal_coordinate}{Horizontal coordinate (azimuth or elevation) to compare to a reference coordinate value or threshold}

\item{horizontal_star_time}{"setting" or "rising". Only required if horizontal_coordinate=="elevation"}

\item{coordinate_threshold}{Horizontal coordinate value used as a reference or threshold for the specified horizonal coordinate of the star during its setting or rising}

\item{visibility_type}{"appearance" (visual magnitude of the sky increasing as days go by) or "disappearance" (visual magnitude of the sky decreasing as days go by)}

\item{silent}{If false a text progress bar is shown}
}
\value{
Time of the star's appearance or disappearance  during its setting or rising at twilight time
}
\description{
Julian day for star's appearance or disappearance
}
\examples{
years<--2000:-1999
from<-str2jd(paste(years[1],"01","01",sep="-"))
to<-str2jd(paste(years[length(years)],"01","01",sep="-"))
star_name<-"epsilon orionis"
stellar_phenomena(from=from,to=to, star_name=star_name,
                  geographic_latitude=str2degrees("42\\u00b036'25\\u0022N"),
                  geographic_longitude=str2degrees("5\\u00b033'58\\u0022W"),
                  geographic_altitude=782,
                  horizontal_coordinate="elevation", horizontal_star_time="setting",
                  coordinate_threshold=2, visibility_type="disappearance", silent=FALSE)
}
