% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{str}
\alias{str}
\title{List formatting to string according to the names of its elements}
\usage{
str(coord)
}
\arguments{
\item{coord}{list of coordinates}
}
\value{
list formatted as string
}
\description{
List formatting to string according to the names of its elements
}
\examples{
coord<-list(declination=c(89.454275227064002,89.5),right_ascension=c(3.8046089729542945,3.9))
str(coord)
}
