% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates.R
\name{str2degrees}
\alias{str2degrees}
\title{Sexagesimal degree parsing from string}
\usage{
str2degrees(str, separators = c("°", "'", "\\""))
}
\arguments{
\item{str}{string dºm's" formatted}

\item{separators}{Separator characters for degrees, minutes and seconds}
}
\value{
Sexagesimal degree
}
\description{
It parses a string into sexagesimal degrees
}
\examples{
str2degrees(c("12\\u00b030'30.25\\u0022","0\\u00b030'30\\u0022"))

}
