% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian_day.R
\name{str2hours}
\alias{str2hours}
\title{Hour parsed from an hh:mm:ss formatted string}
\usage{
str2hours(str)
}
\arguments{
\item{str}{IS8601 formatted string}
}
\value{
Hour angle
}
\description{
Hour parsed from an hh:mm:ss formatted string
}
\examples{
str2hours(c("20:15:30","12:10:05"))

}
