% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian_day.R
\name{str2jd}
\alias{str2jd}
\title{Julian day parsed from an IS8601 formatted string}
\usage{
str2jd(str)
}
\arguments{
\item{str}{IS8601 formatted string}
}
\value{
Julian day
}
\description{
Julian day parsed from an IS8601 formatted string
}
\examples{
str2jd(c("1957-10-04T19:26:024.000","333-01-27T12:00:000.000"))

}
