% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sun.R
\name{true_longitude_sun}
\alias{true_longitude_sun}
\title{True geometric longitude referred to the mean equinox of the date}
\usage{
true_longitude_sun(jd)
}
\arguments{
\item{jd}{Julian day}
}
\value{
Sun's true geometric longitude in degrees
}
\description{
True geometric longitude referred to the mean equinox of the date
}
\examples{
jd<-julian_day(13.19,11,2028)
true_longitude_sun(jd)

}
