% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutation_and_obliquity.R
\name{true_obliquity_ecliptic}
\alias{true_obliquity_ecliptic}
\title{True obliquity of the ecliptic}
\usage{
true_obliquity_ecliptic(jd, precise = F)
}
\arguments{
\item{jd}{Julian day}

\item{precise}{Precise calculation}
}
\value{
True obliquity of the ecliptic
}
\description{
True obliquity of the ecliptic
}
\examples{
mean_obliquity_ecliptic(c(2436116.31, 1842713.00))

}
